backend/requirements.txt
pathlib  
mongoengine
kafka
pydantic 
sockets
fastapi 
typing
passlib
datetime
uvicorn
python-multipart
python-dotenv
python-jose
python-time
bcrypt